move .on event handling into tooltip-on-element.didInsertElement
