+++ addon/components/tooltip-and-popover.js
@@ -225,71 +225,11 @@ export default EmberTetherComponent.extend({
       Ember.assert('You must specify a target attribute in the format target="#element-id" for the tooltip component');
     }
 
-    const event = this.get('event');
     const $target = $(this.get('target'));
     const _tether = this.get('_tether');
     const $_tether = $(_tether.element);
 
     this.sendAction('onTooltipRender', this);
-
-    /* Setup event handling to hide and show the tooltip */
-
-    if (event !== 'none') {
-      const _hideOn = this.get('_hideOn');
-      const _showOn = this.get('_showOn');
-
-      /* If show and hide are the same (e.g. click), toggle
-      the visibility */
-
-      if (_showOn === _hideOn) {
-        $target.on(_showOn, () => {
-          this.toggle();
-        });
-      } else {
-
-        /* Else, add the show and hide events individually */
-
-        if (_showOn !== 'none') {
-          $target.on(_showOn, () => {
-            this.show();
-          });
-        }
-
-        if (_hideOn !== 'none') {
-          $target.on(_hideOn, () => {
-            this.hide();
-          });
-        }
-      }
-
-      /* Hide and show the tooltip on focus and escape
-      for acessibility */
-
-      if (event !== 'focus') {
-
-        /* If the event is click, we don't want the
-        click to also trigger focusin */
-
-        if (event !== 'click') {
-          $target.focusin(() => {
-            this.show();
-          });
-        }
-
-        $target.focusout(() => {
-          this.hide();
-        });
-      }
-
-      $target.keydown((keyEvent) => {
-        if (keyEvent.which === 27) {
-          this.hide();
-          keyEvent.preventDefault();
-
-          return false;
-        }
-      });
-    }
 
     $target.attr({
       'aria-describedby': `#${this.get('elementId')}`,
